# shop-api/docs/swagger.yaml
basePath: /
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      status:
        description: 0=禁用,1=启用
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      detail_images:
        type: string
      gold_price:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      status:
        description: 0=下架,1=上架
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: integer
    type: object
  models.ThirdPartySite:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_alive:
        description: 0=不可用,1=可用
        type: integer
      last_check:
        type: string
      name:
        type: string
      priority:
        type: integer
      status:
        description: 0=禁用,1=启用
        type: integer
      updated_at:
        type: string
      url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 爱影视频APP商城系统和第三方站点管理API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 爱影视频商城API
  version: "1.0"
paths:
...
  /api/v1/shop/products:
    get:
      consumes:
      - application/json
      description: 获取商品列表，支持分页、分类筛选、关键字搜索和排序
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页记录数，默认20
        in: query
        name: limit
        type: integer
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      - description: 排序方式，可选值：price_asc, price_desc
        in: query
        name: sort
        type: string
      - description: 状态，0=下架,1=上架，默认全部
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取商品列表
      tags:
      - 商品
    post:
      consumes:
      - application/json
      description: 创建新商品
      parameters:
      - description: 商品信息
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: 创建商品
      tags:
      - 商品
  /api/v1/shop/products/{id}:
    delete:
      consumes:
      - application/json
      description: 删除商品
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 删除商品
      tags:
      - 商品
    get:
      consumes:
      - application/json
      description: 根据ID获取商品详情
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取商品详情
      tags:
      - 商品
    put:
      consumes:
      - application/json
      description: 更新商品信息
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 商品信息
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 更新商品
      tags:
      - 商品
  /api/v1/sites:
    get:
      consumes:
      - application/json
      description: 获取站点列表，支持分页
      parameters:
      - description: 页码，默认1
        in: query
        name: page
        type: integer
      - description: 每页记录数，默认20
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取站点列表
      tags:
      - 第三方站点
    post:
      consumes:
      - application/json
      description: 创建新站点
      parameters:
      - description: 站点信息
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/models.ThirdPartySite'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      summary: 创建站点
      tags:
      - 第三方站点
  /api/v1/sites/{id}:
    delete:
      consumes:
      - application/json
      description: 删除站点
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 删除站点
      tags:
      - 第三方站点
    get:
      consumes:
      - application/json
      description: 根据ID获取站点详情
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 获取站点详情
      tags:
      - 第三方站点
    put:
      consumes:
      - application/json
      description: 更新站点信息
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 站点信息
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/models.ThirdPartySite'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 更新站点
      tags:
      - 第三方站点
  /api/v1/sites/{id}/check:
    post:
      consumes:
      - application/json
      description: 检测站点是否可访问
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 检测站点健康状态
      tags:
      - 第三方站点
  /api/v1/sites/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新站点启用/禁用状态
      parameters:
      - description: 站点ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态信息，例如：{\
        in: body
        name: status
        required: true
        schema:
          additionalProperties:
            type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
      summary: 更新站点状态
      tags:
      - 第三方站点
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"